% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdfm-package.R
\docType{package}
\name{bdfm-package}
\alias{bdfm-package}
\alias{bdfm}
\title{bdfm: ...}
\description{
\pkg{bdfm} is ....

  The best way to start is to have a look at the vignette:

  \code{vignette("seas")}
}
\references{
Sax C, Eddelbuettel D (2018). "Seasonal Adjustment by X-13ARIMA-SEATS
  in R." \emph{Journal of Statistical Software}, \emph{87}(11), 1-17. doi:
  10.18637/jss.v087.i11 (\url{http://doi.org/10.18637/jss.v087.i11}).
}
\seealso{
\code{\link{dfm}} for the core function and more information on
  package usage.
}
\keyword{package}
