% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstDFM.R
\name{bdfm}
\alias{bdfm}
\title{Estimate Bayesian dynamic factor model}
\usage{
bdfm(Y, factors = 1, lags = 2, forecast = 0, B_prior = NULL,
  lam_B = 0, H_prior = NULL, lam_H = 0, nu_q = 0, nu_r = NULL,
  ID = "PC_full", intercept = TRUE, reps = 1000, burn = 500)
}
\arguments{
\item{Y}{Data in matrix format with time in rows}

\item{factors}{number of factors}

\item{lags}{number of lags in transition equation}

\item{forecast}{number of periods ahead to forecast}

\item{B_prior}{prior matrix for B in the transition equation. Default is zeros.}

\item{lam_B}{prior tightness on B}

\item{H_prior}{prior matrix for H (loadings) in the observation equation. Default is zeros.}

\item{lam_H}{prior tightness on H}

\item{nu_q}{prior deg. of freedom for transition equation, entered as vector with length equal to the number of factors.}

\item{nu_r}{prior deg. of freedom for observables, entered as vector with length equal to the number of observables.}

\item{ID}{Factor Identification. 'PC_full' is the default (using all observed series), 'PC_sub' finds a submatrix of the data that maximizes the number of observations for a square (no missing values) data set. Use 'PC_sub' when many observations are missing.}

\item{intercept}{logical, should an icept be included?}

\item{reps}{number of repetitions for MCMC sampling}

\item{burn}{number of iterations to burn in MCMC sampling}
}
\description{
Estimate Bayesian dynamic factor model
}
