% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdfm-package.R
\docType{data}
\name{econ_us}
\alias{econ_us}
\title{US Economic Data}
\format{Time series of class \code{"ts"}.}
\source{
FRED
}
\description{
US Economic Data
}
\examples{

tail(econ_us)
}
\keyword{datasets}
