% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{End_Last_Month}
\alias{End_Last_Month}
\title{End of Last Month Date}
\usage{
End_Last_Month(last_date)
}
\arguments{
\item{last_date}{date in the month of interest}
}
\description{
Get date for the end of the previous month
}
