% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{End_of_Month}
\alias{End_of_Month}
\title{Get date for last day of month shift months ago or ahead}
\usage{
End_of_Month(last_date, shift = 0)
}
\description{
Get date for last day of month shift months ago or ahead
}
