% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{End_This_Month}
\alias{End_This_Month}
\title{End of Current Month Date}
\usage{
End_This_Month(last_date)
}
\arguments{
\item{last_date}{date in the month of interest}
}
\description{
Get date for the end of the current month
}
