% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm-methods.R
\name{factors}
\alias{factors}
\title{Extractor Functions for Dynamic Factor Models}
\usage{
factors(x)
}
\arguments{
\item{x}{obeject of class `"dfm"`}
}
\description{
Extractor Functions for Dynamic Factor Models
}
