% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstDFM.R
\name{mlDFM}
\alias{mlDFM}
\title{ML Estimation for Dynamic Factor Models}
\usage{
mlDFM(Y, factors = 1, lags = 2, forecast = 0, tol = 0.01,
  Loud = FALSE)
}
\arguments{
\item{Y}{data with time indexed by row}

\item{factors}{number of factors to estimate}

\item{lags}{number of lags in the transition equation}

\item{forecast}{number of periods ahead to forecast}

\item{tol}{tollerance for convergence of the EM Algorithm}

\item{Loud}{T/F, print change in log likelihood at each iteration}
}
\description{
Estimate a dynamic factor model by maximum likelihood using the EM Algorithm in Watson and Engle (1983).
}
