% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale}
\alias{scale}
\title{scale}
\usage{
scale(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{a numeric matrix(like object)}

\item{center}{either a logical value or a numeric vector of length equal to the number of columns of x}

\item{scale}{either a logical value or a numeric vector of length equal to the number of columns of x}
}
\description{
Scale data for factor model estimation and store scaling values
}
