% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seas_factors.R
\name{seas_factors}
\alias{seas_factors}
\alias{dist_unif}
\alias{dist_norm}
\title{Seasonal Factors at Daily, Monthly or Quarterly Frequency}
\usage{
seas_factors(dates, effect = c("months", "workdays"), holiday = easter,
  start = -1, end = 0, dist_fun = NULL)

dist_unif(x)

dist_norm(x)
}
\arguments{
\item{dates}{vector of class `Date`}

\item{effect}{character, one of "months", "quarters", "workdays", "weekdays"}

\item{holiday}{vector of class `Date`, holidays to adjust for, e.g. `easter}

\item{start}{integer, shift start of holiday}

\item{end}{integer, shift end of holiday}

\item{dist_fun}{distribution function for `effect` `"months"` and `"quarters"`. E.g., `dist_norm`.}

\item{x}{numeric vector to be distributed}
}
\description{
Generate seasonal factors at daily, monthly or quarterly frequency
}
\examples{
from <- as.Date("2000-01-01")
to <- as.Date("2005-03-01")
dates.d <- seq(from, to, by = "day")
dates.m <- seq(from, to, by = "month")
dates.q <- seq(from, to, by = "quarter")
seas_factors(dates.d)
seas_factors(dates.m)
seas_factors(dates.q)
seas_factors(dates.d, dist_fun = dist_norm)
}
