% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seas_we.R
\name{seas_we}
\alias{seas_we}
\title{Seasonal adujustment, usign Watson and Engle (1983) Algorithm}
\usage{
seas_we(data, lags = 1, ..., ahead = 0, transformation = "auto",
  span = 0.5, tol = 0.01, verbose = FALSE)
}
\arguments{
\item{data}{data}

\item{lags}{integer, number of lags in transition equation}

\item{...}{further arguments, passed to [seas_factors()]}

\item{ahead}{number of periods ahead to produce adjustments and forecasts}

\item{transformation}{string, transform data before calculating seasonality ("auto", "log", or "none")}

\item{span}{numeric, span for loess() reg used to detrend non-stationary series. NULL implies do not detrend.}

\item{tol}{numeric, tolerance for convergence of likelihood function (times 100)}

\item{verbose}{logical, whether to output convergence of iterations}
}
\value{
an object of class `"seas_we"`. Use `predict` to extract the seasonally adjusted values.
}
\description{
to produce maximum likelihood estimates of model parameters
}
\examples{
m <- seas_we(mdeaths)
adj_we <- predict(m)
library(seasonal)
adj_x13 <- predict(seas(mdeaths))
tsbox::ts_plot(mdeaths, adj_x13, adj_we)
}
