% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA_Programs.R
\name{SeasAdj_WE}
\alias{SeasAdj_WE}
\title{Watson and Engle (1983) Algorithm}
\usage{
SeasAdj_WE(y, N, p = 1, tol = 0.01, Loud = FALSE)
}
\arguments{
\item{y}{data}

\item{N}{matrix of predetermined factors}

\item{p}{lags in transition equation}

\item{tol}{tolerance for convergence of likelihood function (*100)}

\item{Loud}{T/F whether to output convergence of iterations}
}
\description{
to produce maximum likelihood estimates of model parameters
}
