% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_process.R
\name{should_log_diff}
\alias{should_log_diff}
\title{Automatically determine if time series data should be differenced or log differenced}
\usage{
should_log_diff(X)
}
\arguments{
\item{X}{data with time indexed by rows}
}
\description{
Automatically determine if time series data should be differenced or log differenced
}
\examples{
should_log_diff(EuStockMarkets)
}
