% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimData.R
\name{SimDFM}
\alias{SimDFM}
\title{Simulate Dynamic Factor Model}
\usage{
SimDFM(r)
}
\arguments{
\item{r}{number of rows, i.e. time periods (T is TURE in R, but I use T in
C++)}
}
\description{
Simulate Dynamic Factor Model
}
