% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unscale}
\alias{unscale}
\title{unscale}
\usage{
unscale(x, idx = NULL)
}
\arguments{
\item{x}{a numeric matrix(like object)}

\item{idx}{An index of which elements of scaled_center and scaled_scale to use}
}
\description{
Unscale data using values from bdfm::scale
}
