% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdfm-package.R
\docType{package}
\name{bdfm-package}
\alias{bdfm-package}
\alias{bdfm}
\title{Bayesain and Maximum Likelihood Estimation of Dynamic Factor Models}
\description{
Estimates dynamic factor models by simulation using the Durbin and Koopman
(2012) disturbance smoother. Maximum likelihood estimation via Watson and
Engle (1983) and 2-step estimation via principal components is also
supported. Input data may be mixed frequency, noisy, have missing values, or
ragged edges with different start or end dates.
}
\details{
The best way to start is to have a look at the vignette:

\code{vignette("dfm")}
}
\seealso{
\code{\link{dfm}} for the core function and more information on
package usage.
}
\keyword{package}
