% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstDFM.R
\name{BDFM}
\alias{BDFM}
\title{Estimate uniform frequency dynamic factor model}
\usage{
BDFM(Y, m, p, Bp = NULL, lam_B = 1, Hp = NULL, lam_H = 1,
  nu_q = 1, nu_r = 1, ID = "PC_full", ITC = T, reps = 1000)
}
\arguments{
\item{Y}{Data in matrix format with time in rows}

\item{m}{number of factors m}

\item{lam_B}{prior tightness on B (Bp, the prior for B in the tranistion eq. is built in as zero)}

\item{lam_H}{prior tightness on H (additive. A value equal to the number of time periods in Data is a pretty strong prior.)}

\item{nu_q}{prior deg. of freedom for transition equation}

\item{nu_r}{prior deg. of freedom for observables used to identify the model}

\item{ID}{Factor Identification. The default is to use principal components.}

\item{reps}{number of repetitions from which to sample}
}
\description{
Estimate uniform frequency dynamic factor model
}
