% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDFM.R
\name{Cppbdfm}
\alias{Cppbdfm}
\title{MCMC Routine for Bayesian Dynamic Factor Models}
\usage{
Cppbdfm(B, Bp, Jb, lam_B, q, nu_q, H, Hp, lam_H, R, nu_r, Y, freq, LD,
  Ystore = FALSE, store_idx = 0, reps = 1000, burn = 500,
  Loud = FALSE)
}
\arguments{
\item{B}{initial guess for B in transition equation}

\item{Bp}{prior for B}

\item{lam_B}{prior tightness for B (additive)}

\item{q}{initial guess for q in the transition equation}

\item{nu_q}{prior "degrees of freedom" for inverse-Whishart prior for q (additive, prior scale is fixed so that increasing nu_q shrinks the variance towards zero)}

\item{H}{initial guess for H in the trasition equation}

\item{Hp}{prior for H}

\item{lam_H}{prior tightness for H (additive)}

\item{R}{initial guess for diagonal elements of R in the transition equation, entered as a vector}

\item{nu_r}{prior deg. of freedom for elements of R, entered as a vector (additive, prior scale is fixed so that increasing nu_r[j] shrinks the variance of shocks to series j towards zero)}

\item{Y}{Input data. Data must be scaled and centered prior to estimation if desired.}

\item{Ystore}{T/F, should the distribution of Y be stored}

\item{store_idx, }{if Ystore is TRUE, index of which observed series to store. Note C++ uses zero indexing (i.e. subtract 1 from the R index value)}

\item{reps}{number of repetitions for MCMC sampling}

\item{burn}{number of iterations to burn in MCMC sampling}

\item{Loud}{print status of function during evaluation.}
}
\description{
\code{Cppbdfm} is the core C++ function for estimating a linear-Gaussian Bayesain dynamic factor model by MCMC methods using Durbin and Koopman's disturbance smoother. This function may be called directly by advanced users. The only dependencies are the  Armadillo (\url{http://arma.sourceforge.net/}) linear algebra library for C++ and the packages needed for interfacing with R (\code{\link{Rcpp}} and \code{\link{RcppArmadillo}}).
}
