% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstDFM.R
\name{dfm}
\alias{dfm}
\title{Estimate Bayesian dynamic factor model}
\usage{
dfm(Y, factors = 1, lags = 2, forecast = 0, method = "Bayesian")
}
\arguments{
\item{Y}{Data in matrix format with time in rows}

\item{factors}{number of factors}

\item{lags}{number of lags in transition equation}

\item{forecast}{number of periods ahead to forecast}
}
\description{
Estimate Bayesian dynamic factor model
}
