% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm.R
\name{dfm}
\alias{dfm}
\title{Estimate dynamic factor model}
\usage{
dfm(Y, factors = 1, lags = 2, forecast = 0, method = "Bayesian",
  B_prior = NULL, lam_B = 0, H_prior = NULL, lam_H = 0, nu_q = 0,
  nu_r = NULL, identification = "PC_full", intercept = TRUE,
  reps = 1000, burn = 500, Loud = F, EM_tolerance = 0.01)
}
\arguments{
\item{Y}{data in matrix format with time in rows}

\item{factors}{number of factors}

\item{lags}{number of lags in transition equation}

\item{forecast}{number of periods ahead to forecast}

\item{method}{character, method to be used}

\item{B_prior}{prior matrix for B in the transition equation. Default is zeros.}

\item{lam_B}{prior tightness on B}

\item{H_prior}{prior matrix for H (loadings) in the observation equation. Default is zeros.}

\item{lam_H}{prior tightness on H}

\item{nu_q}{prior deg. of freedom for transition equation, entered as vector with length equal to the number of factors.}

\item{nu_r}{prior deg. of freedom for observables, entered as vector with length equal to the number of observables.}

\item{identification}{factor identification. 'PC_full' is the default (using all observed series), 'PC_sub' finds a submatrix of the data that maximizes the number of observations for a square (no missing values) data set. Use 'PC_sub' when many observations are missing.}

\item{intercept}{logical, should an icept be included?}

\item{reps}{number of repetitions for MCMC sampling}

\item{burn}{number of iterations to burn in MCMC sampling}

\item{Loud}{print status of function during evalutation. If ML, print difference in likelihood at each iteration of the EM algorithm.}

\item{EM_tolerance}{tolerance for convergence of EM algorithm. Convergence criteria is calculated as 200 * (Lik1 - Lik0) / abs(Lik1 + Lik0) where Lik1 is the log likelihood from this iteration and Lik0 is the likelihood from the previous iteration.}
}
\description{
Estimate dynamic factor model
}
