% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm.R
\name{dfm}
\alias{dfm}
\title{Estimate dynamic factor model}
\usage{
dfm(data, factors = 1, lags = "auto", forecasts = 0,
  method = c("bayesian", "ml", "pc"), scale = TRUE, logs = "auto",
  diffs = "auto", outlier_threshold = 4, frequency_mix = "auto",
  pre_differenced = NULL, trans_prior = NULL, trans_shrink = 0,
  trans_df = 0, obs_prior = NULL, obs_shrink = 0, obs_df = NULL,
  identification = "pc_long", store_idx = NULL, reps = 1000,
  burn = 500, verbose = interactive(), tol = 0.01,
  return_intermediates = FALSE, orthogonal_shocks = FALSE)
}
\arguments{
\item{data}{data in matrix format with time in rows}

\item{factors}{number of factors}

\item{lags}{number of lags in transition equation}

\item{forecasts}{number of periods ahead to forecasts}

\item{method}{character, method to be used}

\item{scale}{scale data before estimation (True/False)?}

\item{logs}{names or index values of series which should be entered in logs (and potentially differenced)}

\item{diffs}{names or index values of series which should be differenced}

\item{outlier_threshold}{drop observations more than x standard deviations from the mean}

\item{frequency_mix}{'auto' or numeric, number of high frequency periods in observation if data is mixed frequency}

\item{pre_differenced, }{names or index values of low freqeuncy series entered in differences (not necessary if already specified in 'diffs')}

\item{trans_prior}{prior matrix for B in the transition equation. Default is
zeros.}

\item{trans_shrink}{prior tightness on B matrix in trasition equation}

\item{trans_df}{prior deg. of freedom for transition equation}

\item{obs_prior}{prior matrix for H (loadings) in the observation equation
Default is zeros.}

\item{obs_shrink}{prior tightness on H (loadings) in the observation equation}

\item{obs_df}{prior deg. of freedom for observables, entered as vector with
length equal to the number of observables.}

\item{identification}{factor identification. 'pc_long' is the default and finds series with the most
observations over time. 'pc_full' uses all observed series, 'pc_sub' finds a submatrix of the data that maximizes
  the number of observations for a square (no missing values) data set. Users may also enter a
  numeric vector for specified series.}

\item{store_idx}{if estimation is Bayesian, index of input data to store the full posterior distribution of predicted values.}

\item{reps}{number of repetitions for MCMC sampling}

\item{burn}{number of iterations to burn in MCMC sampling}

\item{verbose}{print status of function during evalutation. If ML, print
difference in likelihood at each iteration of the EM algorithm. Default is `TRUE` in interactive mode, `FALSE` otherwise.}

\item{tol}{tolerance for convergence of EM algorithm (method `ml` only). Convergence
criteria is calculated as 200 * (Lik1 - Lik0) / abs(Lik1 + Lik0) where Lik1
is the log likelihood from this iteration and Lik0 is the likelihood from
the previous iteration.}

\item{return_intermediates}{if data is mixed frequency, should estimation return intermediate values of low frequency variables}

\item{orthogonal_shocks}{Return a rotation of the model with orthogonal shocks and factors}
}
\description{
Estimate dynamic factor model
}
\examples{
fdeaths0 <- fdeaths
fdeaths0[length(fdeaths0)] <- NA
dta <- cbind(fdeaths0, mdeaths)

library(bdfm)
m <- dfm(dta, forecasts = 2)
summary(m)
}
