% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{End_of_Month}
\alias{End_of_Month}
\title{End of Month Date}
\usage{
End_of_Month(last_date, shift = 0)
}
\arguments{
\item{last_date}{date in the month of interest}

\item{shift}{Months ahead (positive value) or behind (negative value)}
}
\description{
Get date for the last day of month shift months ago (-) or ahead (+)
}
