% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfm-methods.R
\name{factors}
\alias{factors}
\alias{adjusted}
\alias{predict.dfm}
\title{Extractor Functions for Dynamic Factor Models}
\usage{
factors(object)

adjusted(object)

\method{predict}{dfm}(object, ...)
}
\arguments{
\item{object}{object of class \code{"dfm"}}

\item{...}{unused, to comply with generic}
}
\description{
\code{predict} extracts the predicted values of model. \code{adjusted} returns the
original series with the predicted values substituted if missing. \code{factors}
return the factor(s) of the model.
}
\examples{
\dontrun{
# predict vs. adjusted
library(tsbox)
dta0 <- ts_seas(cbind(mdeaths, fdeaths))  # seasonally adjust
dta <- dta0
dta[1:10, 2] <- NA

m <- dfm(dta)
ts_plot(predict(m)[, 'fdeaths'], dta0[, 'fdeaths'])
ts_plot(adjusted(m)[, 'fdeaths'], dta0[, 'fdeaths'])
}
}
