% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA_Programs.R
\name{Predetermined.m}
\alias{Predetermined.m}
\title{Seasonal Factors at Monthly Frequency}
\usage{
Predetermined.m(dates, predetermined)
}
\arguments{
\item{dates}{dates that factors will be generated for (date vector)}

\item{predetermined}{seasonal adjustment factors to be generated, for example c('June', 'July')}
}
\description{
Generate predetermined seasonal factors at monthly frequency
}
\examples{
Spain <- read_csv(system.file("Examples/Spain_IP.csv", package = "BDFM"),
  col_types = cols(date = col_date(format = "\%Y-\%m-\%d"))
)
y <- Spain$IP
dates <- Spain$date
N <- Predetermined.m(dates, predetermined = c("January", "February", "March", "April", "May", "June", "July", "August", "September", "October", "November", "December", "Easter", "trading_days"))
}
