% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA_Programs.R
\name{Predetermined.m}
\alias{Predetermined.m}
\title{Seasonal Factors at Monthly Frequency}
\usage{
Predetermined.m(dates, predetermined)
}
\arguments{
\item{dates}{dates that factors will be generated for (date vector)}

\item{predetermined}{seasonal adjustment factors to be generated, for example c('June', 'July')}
}
\description{
Generate predetermined seasonal factors at monthly frequency
}
